Feature/AX-7147 - Private Saved Queries (#5602)

+++ plugins/gui/routes/entities/views/views_generator.py
@@ -2,7 +2,7 @@ import re
 from flask import (jsonify,
                    request)
 
-from axonius.plugin_base import EntityType
+from axonius.plugin_base import EntityType, return_error
 from axonius.utils.gui_helpers import (paginated, filtered, sorted_endpoint)
 from axonius.utils.permissions_helper import PermissionCategory, PermissionAction, PermissionValue
 from axonius.consts.gui_consts import ACTIVITY_PARAMS_COUNT
@@ -29,26 +29,52 @@ def views_generator(base_permission_category: PermissionCategory):
             return jsonify(
                 self._get_entity_views(self.entity_type, limit, skip, mongo_filter, mongo_sort, query_type))
 
-        @gui_route_logged_in(methods=['PUT'])
-        def add_entity_views(self):
+        @gui_route_logged_in(methods=['PUT'], proceed_and_set_access=True)
+        def add_entity_views(self, no_access):
             """
             Add/ Update views over the devices db
             :return:
             """
+            # If there are no permissions, also check that the request doesn't relate to the private query
+            # Only then we report error and exit
+            if no_access and not request.get_json().get('private'):
+                return return_error('You are lacking some permissions for this request', 401)
+
             return jsonify(self._add_entity_view(self.entity_type))
 
         @filtered()
-        @gui_route_logged_in('<query_type>', methods=['DELETE'], activity_params=[ACTIVITY_PARAMS_COUNT])
-        def delete_views(self, query_type, mongo_filter):
+        @gui_route_logged_in('<query_type>', methods=['DELETE'], activity_params=[ACTIVITY_PARAMS_COUNT],
+                             proceed_and_set_access=True)
+        def delete_views(self, query_type, mongo_filter, no_access):
             """
             Delete views over the devices db
             """
+            # If there are no permissions, also check that the request doesn't relate to the private query
+            # Only then we report error and exit
+            if no_access and not request.get_json().get('private'):
+                return return_error('You are lacking some permissions for this request', 401)
+
             return jsonify({ACTIVITY_PARAMS_COUNT: str(self._delete_entity_views(self.entity_type, mongo_filter))})
 
-        @gui_route_logged_in('<query_id>', methods=['POST'])
-        def views_update(self, query_id):
+        @gui_route_logged_in('<query_id>', methods=['POST'], proceed_and_set_access=True)
+        def views_update(self, query_id, no_access):
             """
             Update name of an existing view
+            :return:
+            """
+            # If there are no permissions, also check that the request doesn't relate to the private query
+            # Only then we report error and exit
+            if no_access and not request.get_json().get('private'):
+                return return_error('You are lacking some permissions for this request', 401)
+
+            self._update_entity_views(self.entity_type, query_id)
+            return ''
+
+        @gui_route_logged_in('<query_id>/publish', methods=['POST'], required_permission=PermissionValue.get(
+            PermissionAction.Add, base_permission_category, PermissionCategory.SavedQueries))
+        def views_publish(self, query_id):
+            """
+            Sets private view to public
             :return:
             """
             self._update_entity_views(self.entity_type, query_id)
